+++ lib/flutter_scene.dart
@@ -1,4 +1,0 @@
-library flutter_scene;
-
-export 'camera.dart';
-export 'scene.dart';