+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 * Rewrite for Flutter GPU.
 * Physically based rendering.
 * More conventional interface for scene construction.
+
+## 0.1.1
+
+* Rename PhysicallyBasedMaterial and UnlitMaterial
+* Fix environment lighting problems in PhysicallyBasedMaterial.
+* Add default environment map.